ref(explore): Lift explore timeseries query to parent component
There's some dependencies between componentst that causes re-renders. Lift the
timeseries query into the parent to avoid this.
