+++ static/app/views/explore/content.tsx
@@ -1,4 +1,4 @@
-import {useCallback, useState} from 'react';
+import {useCallback, useMemo, useState} from 'react';
 import styled from '@emotion/styled';
 import * as Sentry from '@sentry/react';
 import type {Location} from 'history';
@@ -22,7 +22,8 @@ import {
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
 import {t} from 'sentry/locale';
 import {space} from 'sentry/styles/space';
-import type {Confidence} from 'sentry/types/organization';
+import {defined} from 'sentry/utils';
+import {dedupeArray} from 'sentry/utils/dedupeArray';
 import {DiscoverDatasets} from 'sentry/utils/discover/types';
 import {
   type AggregationKey,
@@ -38,6 +39,7 @@ import {
   useExploreDataset,
   useExploreMode,
   useExploreQuery,
+  useExploreVisualizes,
   useSetExploreQuery,
 } from 'sentry/views/explore/contexts/pageParamsContext';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
@@ -45,8 +47,10 @@ import {
   SpanTagsProvider,
   useSpanTags,
 } from 'sentry/views/explore/contexts/spanTagsContext';
+import {useExploreTimeseries} from 'sentry/views/explore/hooks/useExploreTimeseries';
 import {ExploreTables} from 'sentry/views/explore/tables';
 import {ExploreToolbar} from 'sentry/views/explore/toolbar';
+import {combineConfidenceForSeries} from 'sentry/views/explore/utils';
 
 interface ExploreContentProps {
   location: Location;
@@ -59,6 +63,7 @@ function ExploreContentImpl({}: ExploreContentProps) {
   const {selection} = usePageFilters();
   const dataset = useExploreDataset();
   const mode = useExploreMode();
+  const visualizes = useExploreVisualizes();
 
   const numberTags = useSpanTags('number');
   const stringTags = useSpanTags('string');
@@ -80,11 +85,24 @@ function ExploreContentImpl({}: ExploreContentProps) {
     });
   }, [location, navigate]);
 
-  const [confidences, setConfidences] = useState<Confidence[]>([]);
-  const [chartError, setChartError] = useState<string>('');
   const [tableError, setTableError] = useState<string>('');
 
   const maxPickableDays = 7;
+
+  const {timeseriesResult, canUsePreviousResults} = useExploreTimeseries({query});
+  const confidences = useMemo(
+    () =>
+      visualizes.map(visualize => {
+        const dedupedYAxes = dedupeArray(visualize.yAxes);
+        const series = dedupedYAxes
+          .flatMap(yAxis => timeseriesResult.data[yAxis])
+          .filter(defined);
+        return combineConfidenceForSeries(series);
+      }),
+    [timeseriesResult.data, visualizes]
+  );
+
+  const chartError = timeseriesResult.error?.message ?? '';
 
   return (
     <SentryDocumentTitle title={t('Traces')} orgSlug={organization.slug}>
@@ -173,9 +191,10 @@ function ExploreContentImpl({}: ExploreContentProps) {
                 </Alert>
               )}
               <ExploreCharts
+                canUsePreviousResults={canUsePreviousResults}
+                confidences={confidences}
                 query={query}
-                setConfidences={setConfidences}
-                setError={setChartError}
+                timeseriesResult={timeseriesResult}
               />
               <ExploreTables confidences={confidences} setError={setTableError} />
             </MainSection>
